maintain quoted strings in object literals
It may be valuable to be able to tell the difference between these two
constructs even though their meanings are identical.
{'foo': 1}
and
{foo: 1}

For instance there are some very powerful minifications that can be made
with careful site-wide passes. Also if you are using uglify's AST for
linting this comes in handy.